Remove unnecessary clean up of file in test.
The specific functionality that's causing the symptom noted in PT # 79728226 was introduced via commit https://github.com/fcrepo4/fcrepo4/commit/068c3460e7c302caba4e36a8abc216d02704fee2

Rationale for the PR:

ReadOnlyExternalPropertiesFedoraFileSystemConnectorIT fails when trying to delete a file generated when running
BasicReadWriteFedoraFileSystemConnectorIT. The former makes no use of this file, and hence shouldn't try to delete it (indirectly via AbstractFedoraFileSystemConnectorIT).

This fixes the specific build error that result from attempting to delete a file that's still held on by another Java process. The file lock was observed on Windows Java 7u67 (and 7u25). I used the Windows handle command to see that the file was indeed being held open by a java process. I was unable to get further details about which thread was holding the file open. As this seems to be a common problem on Windows, it's not clear to me whether this is worth investigating.