Implement 4.7.x permissions in the WebAC filter
Require acl:read for GET requests and acl:write mode for all other request methods. Added a unit test for the WebACFilter that creates a mock user and request environment and ensures that the correct HTTP response code (currently, either 200 or 403) is returned. Added an equals() method to the WebACPermission class to make mocking easier, and it allows the intent of the implies() method to be read more cleanly.

Fixes: https://jira.duraspace.org/browse/FCREPO-2762