Fix last modified date tests
On systems with course-grained clocks, subsequent invocations of
Instant.now() may produce the same value.  Two tests compare an
"original" modification date (provided by a user?) with a "last modified
date" (the value that is persisted), and check that the "last modified"
is after the "original" last modified date.  The semantics of these
dates are unclear.  Nevertheless, changing the test from original < last
to original <= last causes the test to pass.