Enable WIndows build (#1634)
* Use forward slash instead of file separator constant
* Enable Windows build in Travis
* Throw if index file cleanup fails
* Fix last modified date tests

On systems with course-grained clocks, subsequent invocations of
Instant.now() may produce the same value.  Two tests compare an
"original" modification date (provided by a user?) with a "last modified
date" (the value that is persisted), and check that the "last modified"
is after the "original" last modified date.  The semantics of these
dates are unclear.  Nevertheless, changing the test from original < last
to original <= last causes the test to pass.

* Use monotonically increasing session creation time

As part of OcflPersistentStorageSession.commit(), sessions are
ordered by their timestamp before they are committed.  The comments
in the code claim this supports testing.  On systems with course
granularity clocks, equal creation times are common.  Since
tests presume a well-ordering of sessions based on sorting based on
time, they should be using monotonically increasing creation times
rather than being at the whim of Instant.now()

* Catch exception when parsing link

Resolves: https://jira.lyrasis.org/browse/FCREPO-3236l